pyproject.toml
[tool.poetry]
name = "wokrout-builder"
version = "0.1.0"
description = ""
authors = ["ytouate <touateyoussef2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
sqlalchemy = "^2.0.25"
python-dotenv = "^1.0.0"
openai = "^1.7.2"
tortoise-orm = {extras = ["psycopg"], version = "^0.20.0"}
jinja2 = "^3.1.3"
pydantic = {extras = ["email"], version = "^2.5.3"}
fastapi-mail = "^1.4.1"
pyjwt = "^2.8.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
